+++ src/server/graphql/blockchain/config.ts
@@ -371,8 +371,9 @@ export function compileCollectionConfig<TItem>(
             // maxJoinDepth is because in nested joins all paths should be different
             const pathForQuery = `${join.pathForQuery}${maxJoinDepth - 1}`
             const params = new QParams({
-                disableKeyComparison:
-                    context.services.config.queries.filter.disableKeyComparison,
+                stringifyKeyInAqlComparison:
+                    context.services.config.queries.filter
+                        .stringifyKeyInAqlComparison,
             })
             const returnExpression = config[
                 join.joinedCollection